refactor(api): move pe usecase and repository in to prescription
