+++ api/src/prescription/campaign-participation/infrastructure/repositories/pole-emploi-sending-repository.js
@@ -1,6 +1,6 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { config } from '../../../lib/config.js';
-import * as OidcIdentityProviders from '../../../src/identity-access-management/domain/constants/oidc-identity-providers.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { config } from '../../../../../lib/config.js';
+import * as OidcIdentityProviders from '../../../../identity-access-management/domain/constants/oidc-identity-providers.js';
 
 const create = function ({ poleEmploiSending }) {
   return knex('pole-emploi-sendings').insert({ ...poleEmploiSending });
